docs(repo): document kebab-case file naming; rename Vectorize provider files
Codifies the existing kebab-case convention so it covers source files,
not just project documentation. Adds an explicit short exceptions list
in CLAUDE.md (React / Astro components, TanStack Router file-based
routes, generated files) and points the top-level AGENTS.md at it.

Renames the in-repo Alchemy v2 Vectorize provider to match:

  src/alchemy/Vectorize/VectorizeIndex.ts        -> vectorize/vectorize-index.ts
  src/alchemy/Vectorize/VectorizeIndexBinding.ts -> vectorize/vectorize-index-binding.ts
  src/alchemy/Vectorize/Providers.ts             -> vectorize/providers.ts
  src/alchemy/Vectorize/index.ts                 -> vectorize/index.ts

Updates internal imports, alchemy.run.ts, the app AGENTS.md, and the
project progress note to the new paths. pnpm typecheck and pnpm build
both pass.
