--- apps/f311x/src/alchemy/Vectorize/index.ts
@@ -1,3 +1,0 @@
-export * from './VectorizeIndex.ts'
-export * from './VectorizeIndexBinding.ts'
-export * from './Providers.ts'
