+++ apps/f311x/src/alchemy/vectorize/index.ts
@@ -1,0 +1,3 @@
+export * from './vectorize-index.ts'
+export * from './vectorize-index-binding.ts'
+export * from './providers.ts'
