+++ docs/projects/f311x/2026-05-16-progress.md
@@ -66,11 +66,11 @@ PRD package / symbol names that did not match what's actually shipped:
 
 ## Vectorize custom resource
 
-Per the guide at https://v2.alchemy.run/guides/custom-provider/, wrote an in-repo custom Alchemy v2 provider for Vectorize at `apps/f311x/src/alchemy/Vectorize/`:
+Per the guide at https://v2.alchemy.run/guides/custom-provider/, wrote an in-repo custom Alchemy v2 provider for Vectorize at `apps/f311x/src/alchemy/vectorize/`:
 
-- `VectorizeIndex.ts` -- Resource declaration (`F311x.Cloudflare.VectorizeIndex`) + `Provider.effect(...)` with `reconcile` (observe -> ensure -> return), `delete` (idempotent, swallows `NotFound`), `diff` (any prop drift -> `replace`, since Vectorize indexes are fully immutable post-create), and `read` (cached or list-scan fallback).
-- `VectorizeIndexBinding.ts` -- `Binding.Service` + `Binding.Policy` mirroring the `KVNamespace` / `Hyperdrive` patterns. Exposes a typed Effect-flavored `Vectorize` client (`describe` / `query` / `insert` / `upsert` / `deleteByIds` / `getByIds`) plus a `raw` escape hatch.
-- `Providers.ts` -- `Provider.collection([VectorizeIndex, VectorizeIndexBindingPolicy])` layered with the live impls; `alchemy.run.ts` merges it onto `Cloudflare.providers()`.
+- `vectorize-index.ts` -- Resource declaration (`F311x.Cloudflare.VectorizeIndex`) + `Provider.effect(...)` with `reconcile` (observe -> ensure -> return), `delete` (idempotent, swallows `NotFound`), `diff` (any prop drift -> `replace`, since Vectorize indexes are fully immutable post-create), and `read` (cached or list-scan fallback).
+- `vectorize-index-binding.ts` -- `Binding.Service` + `Binding.Policy` mirroring the `KVNamespace` / `Hyperdrive` patterns. Exposes a typed Effect-flavored `Vectorize` client (`describe` / `query` / `insert` / `upsert` / `deleteByIds` / `getByIds`) plus a `raw` escape hatch.
+- `providers.ts` -- `Provider.collection([VectorizeIndex, VectorizeIndexBindingPolicy])` layered with the live impls; `alchemy.run.ts` merges it onto `Cloudflare.providers()`.
 - Wired into `alchemy.run.ts` as `Vectorize.VectorizeIndex('Knowledge', { config: { dimensions: 768, metric: 'cosine' } })` and bound into the worker as `KNOWLEDGE`.
 - Added `@distilled.cloud/cloudflare` as a direct dep so the underlying typed REST SDK is resolvable (`createIndex` / `getIndex` / `listIndexes` / `deleteIndex`).
 - `src/alchemy/**` is excluded from `tsc` (same way `alchemy.run.ts` is) because Alchemy v2 peer-deps `effect@>=4.0.0-beta.66` and the rest of the app is on `effect@3.21.2`. The runtime contract is unaffected -- pnpm resolved the peer satisfactorily and the Alchemy CLI loads the file via its own bundler -- but `tsc` can't reconcile the v3-vs-v4 Effect type API. Once we move the app to `effect@4`, drop the exclude.
